refactor: removes AtomicSwapWebSocket
