+++ __tests__/new/swapConnection.test.ts
@@ -1,16 +1,16 @@
 import { AtomicSwapServiceConnection } from '../../src/new/swapConnection';
 import { ConnectionState } from '../../src/wallet/types';
-import AtomicSwapWebSocket from '../../src/websocket/atomic-swap';
+import WalletWebSocket from '../../src/websocket/index';
 
 const atomicSwapServiceWs = 'http://localhost:3002' // mock value
 
 let sendMessageSpy, setupSpy;
 
 beforeAll(() => {
-  sendMessageSpy = jest.spyOn(AtomicSwapWebSocket.prototype, 'sendMessage')
+  sendMessageSpy = jest.spyOn(WalletWebSocket.prototype, 'sendMessage')
     .mockImplementation(jest.fn());
 
-  setupSpy = jest.spyOn(AtomicSwapWebSocket.prototype, 'setup')
+  setupSpy = jest.spyOn(WalletWebSocket.prototype, 'setup')
     .mockImplementation(jest.fn());
 })
